requirements.txt
# NY Open Futures Terminal - Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
plotly>=5.18.0
yfinance>=0.2.33
pydantic>=2.5.0
PyYAML>=6.0.1
python-dateutil>=2.8.2
pytz>=2024.1
scipy>=1.11.0
